# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Easy Agra API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  super_admin: 'super_admin',
  admin: 'admin',
  hotel_owner: 'hotel_owner',
  restaurant_owner: 'restaurant_owner',
  spa_owner: 'spa_owner',
  customer: 'customer',
} as const;

export type UserStatus = typeof UserStatus[keyof typeof UserStatus];


export const UserStatus = {
  pending: 'pending',
  active: 'active',
  suspended: 'suspended',
  rejected: 'rejected',
} as const;

export interface User {
  id: number;
  fullName: string;
  email: string;
  mobile: string;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  role: UserRole;
  status: UserStatus;
  /** @nullable */
  profilePhoto?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface UsersPage {
  users: User[];
  total: number;
  page: number;
  limit: number;
}

export interface RegisterInput {
  fullName: string;
  email: string;
  mobile: string;
  password: string;
  city?: string;
  state?: string;
}

export interface LoginInput {
  email: string;
  password: string;
  rememberMe?: boolean;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface ForgotPasswordInput {
  email: string;
}

export interface ResetPasswordInput {
  token: string;
  password: string;
}

export interface UserUpdate {
  fullName?: string;
  mobile?: string;
  city?: string;
  state?: string;
}

export type UserStatusUpdateStatus = typeof UserStatusUpdateStatus[keyof typeof UserStatusUpdateStatus];


export const UserStatusUpdateStatus = {
  pending: 'pending',
  active: 'active',
  suspended: 'suspended',
  rejected: 'rejected',
} as const;

export interface UserStatusUpdate {
  status: UserStatusUpdateStatus;
  reason?: string;
}

export type UserRoleUpdateRole = typeof UserRoleUpdateRole[keyof typeof UserRoleUpdateRole];


export const UserRoleUpdateRole = {
  super_admin: 'super_admin',
  admin: 'admin',
  hotel_owner: 'hotel_owner',
  restaurant_owner: 'restaurant_owner',
  spa_owner: 'spa_owner',
  customer: 'customer',
} as const;

export interface UserRoleUpdate {
  role: UserRoleUpdateRole;
}

export interface ChangePasswordInput {
  currentPassword: string;
  newPassword: string;
}

export interface PhotoUploadInput {
  photoUrl: string;
}

export type OwnerRequestRequestedRole = typeof OwnerRequestRequestedRole[keyof typeof OwnerRequestRequestedRole];


export const OwnerRequestRequestedRole = {
  hotel_owner: 'hotel_owner',
  restaurant_owner: 'restaurant_owner',
  spa_owner: 'spa_owner',
} as const;

export type OwnerRequestStatus = typeof OwnerRequestStatus[keyof typeof OwnerRequestStatus];


export const OwnerRequestStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface OwnerRequest {
  id: number;
  userId: number;
  requestedRole: OwnerRequestRequestedRole;
  status: OwnerRequestStatus;
  /** @nullable */
  businessName?: string | null;
  /** @nullable */
  businessDescription?: string | null;
  /** @nullable */
  rejectionReason?: string | null;
  /** @nullable */
  reviewedBy?: number | null;
  createdAt: string;
  updatedAt: string;
  user: User;
}

export interface OwnerRequestsPage {
  requests: OwnerRequest[];
  total: number;
  page: number;
  limit: number;
}

export type OwnerRequestInputRequestedRole = typeof OwnerRequestInputRequestedRole[keyof typeof OwnerRequestInputRequestedRole];


export const OwnerRequestInputRequestedRole = {
  hotel_owner: 'hotel_owner',
  restaurant_owner: 'restaurant_owner',
  spa_owner: 'spa_owner',
} as const;

export interface OwnerRequestInput {
  requestedRole: OwnerRequestInputRequestedRole;
  businessName?: string;
  businessDescription?: string;
}

export interface OwnerRequestRejectInput {
  reason: string;
}

export type NotificationType = typeof NotificationType[keyof typeof NotificationType];


export const NotificationType = {
  welcome: 'welcome',
  owner_approved: 'owner_approved',
  owner_rejected: 'owner_rejected',
  account_update: 'account_update',
  announcement: 'announcement',
  general: 'general',
} as const;

export interface Notification {
  id: number;
  userId: number;
  title: string;
  message: string;
  type: NotificationType;
  isRead: boolean;
  createdAt: string;
}

export interface NotificationsPage {
  notifications: Notification[];
  total: number;
  page: number;
  limit: number;
}

export interface UnreadCount {
  count: number;
}

/**
 * @nullable
 */
export type ActivityLogUser = {
  id?: number;
  fullName?: string;
  email?: string;
} | null;

export interface ActivityLog {
  id: number;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  userRole?: string | null;
  actionType: string;
  description: string;
  /** @nullable */
  ipAddress?: string | null;
  createdAt: string;
  /** @nullable */
  user?: ActivityLogUser;
}

export interface ActivityLogsPage {
  logs: ActivityLog[];
  total: number;
  page: number;
  limit: number;
}

export interface PlatformSettings {
  id: number;
  appName: string;
  /** @nullable */
  logo?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  supportEmail?: string | null;
  /** @nullable */
  termsAndConditions?: string | null;
  /** @nullable */
  privacyPolicy?: string | null;
  maintenanceMode: boolean;
  updatedAt: string;
}

export interface PlatformSettingsUpdate {
  appName?: string;
  logo?: string;
  contactEmail?: string;
  supportEmail?: string;
  termsAndConditions?: string;
  privacyPolicy?: string;
  maintenanceMode?: boolean;
}

export interface DashboardStats {
  totalUsers: number;
  activeUsers: number;
  pendingUsers: number;
  suspendedUsers: number;
  totalOwnerRequests: number;
  pendingOwnerRequests: number;
  totalAdmins: number;
}

export interface RoleCount {
  role: string;
  count: number;
}

export interface AnnouncementInput {
  title: string;
  message: string;
}

export type ListUsersParams = {
role?: string;
status?: string;
search?: string;
page?: number;
limit?: number;
};

export type ListOwnerRequestsParams = {
status?: string;
page?: number;
limit?: number;
};

export type ListNotificationsParams = {
unreadOnly?: boolean;
page?: number;
limit?: number;
};

export type ListActivityLogsParams = {
userId?: number;
actionType?: string;
page?: number;
limit?: number;
};