# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Easy Agra API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';
import type { UserStatus } from './userStatus';

export interface User {
  id: number;
  fullName: string;
  email: string;
  mobile: string;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  role: UserRole;
  status: UserStatus;
  /** @nullable */
  profilePhoto?: string | null;
  createdAt: string;
  updatedAt: string;
}